% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_op.R
\name{cache_info}
\alias{cache_info}
\title{Get info about the cache}
\usage{
cache_info(summary_only = FALSE)
}
\arguments{
\item{summary_only}{Should we only print summary info?}
}
\value{
info about files in the cache
}
\description{
Returns table with hashes, size of files (in MB), when each file was last modified, and the age in days of each file.
}
\examples{
cache_info(FALSE)
}

