% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_op.R
\name{cache_op}
\alias{cache_op}
\title{Wrapper to add caching to functions}
\usage{
cache_op(fun, query, use_cache = T, ...)
}
\arguments{
\item{fun}{function name (string)}

\item{query}{first argument to fun}

\item{use_cache}{should we use cache this time}

\item{...}{other parameters passed to fun}
}
\value{
typical fun, query, ellipsis result
}
\description{
Wrapper to add caching to functions
}
\examples{
cache_op("sin", pi)
}

