% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catcheR_10Xcatch.R
\name{catcheR_10Xcatch}
\alias{catcheR_10Xcatch}
\title{catcheR_10Xcatch}
\usage{
catcheR_10Xcatch(
  group = c("docker", "sudo"),
  folder,
  fastq.read1,
  fastq.read2,
  expression.matrix,
  reference = "GGCGCGTTCATCTGGGGGAGCCG",
  UCI.length = 6,
  threads = 2,
  percentage = 15,
  mode = "bimodal",
  ratio = 5,
  samples = 1,
  x = 100,
  y = 400
)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{folder, }{a character string indicating the path of the working folder containing the input files}

\item{fastq.read1, }{a character string indicating the filename of read 1 fastq or fastq.gz containing barcodes sequencing}

\item{fastq.read2, }{a character string indicating the filename of read 2 fastq or fastq.gz containing barcodes sequencing}

\item{expression.matrix, }{a character string indicating the filename of the gene expression matrix csv}

\item{reference, }{a character string indicating the sequence to identify reads containing barcodes. Should be found at at the beginning of read2. Use reverse complement!}

\item{threads, }{integer number of threads to be used for parallelization}

\item{percentage, }{integer threshold of percentage of UMIs supporting a UCI over total UMIs supporting UCIs in the same cell, to consider the UCI valid. Suggested default is 15.}

\item{mode, }{a character string. Two options: "bimodal" or "noise". To evaluate a threshold number of UMIs to consider a UCI valid there are 2 options: "bimodal" (default) which sets the threshold at the valley of the UMIxUCI distribution, or "noise", which sets the threshold at 1.35 * number of UCI supported by a single UMI.}

\item{ratio, }{select valid cells = only 1 true UCI and top UCI UMIs / second top UCI UMIs > ratio}

\item{samples, }{integer indicating the number of different samples present in the sample (aggregated with cell ranger aggr)}

\item{x, }{an integer indicating the x limit to zoom the UMIxUCI plot. Default is 100}

\item{y, }{an integer indicating the y limit to zoom the UMIxUCI plot. Default is 400}

\item{UCI.legth, }{integer indicating the length of unique clonal identifier. Should be found after reference on read2}
}
\value{
a gene expression matrix where cell names contains annotations about the perturbation present in each cell + QC plots
}
\description{
Data analysis pipeline for the iPS2-seq methods from HEDGe lab.
}
\examples{

folder = "/20tb/ratto/catcheR/test_02_7/"
catcheR_10Xcatch(group = "docker", 
                folder = folder, 
                fastq.read1 = list.files(folder, pattern = "R1"), 
                fastq.read2 = list.files(folder, pattern = "R2"), 
                expression.matrix = "matrix.csv", 
                reference = "GGCGCGTTCATCTGGGGGAGCCG",
                UCI.length = 6,
                threads = 12, 
                percentage = 15,
                mode = "noise",
                ratio = 5,
                samples = 4)

}
\author{
Maria Luisa Ratto, marialuisa.ratto [at] unito [dot] it, UNITO
}
