% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catcheR_10Xnocatch.R
\name{catcheR_10Xnocatch}
\alias{catcheR_10Xnocatch}
\title{catcheR_10Xnocatch}
\usage{
catcheR_10Xnocatch(
  group = c("docker", "sudo"),
  folder,
  expression.matrix,
  threshold,
  sample = 1,
  reference = "TACGCGTTCATCTGGGGGAGCCG"
)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{folder, }{a character string indicating the path of the working folder containing the input files}

\item{expression.matrix, }{a character string indicating the filename of the gene expression matrix csv}

\item{threshold, }{an integer indicating the minimum number of UMIs associated to the empty reference to consider a cell empty}

\item{sample, }{integer indicating the sample to analyze (between those aggregated with cell ranger aggr)}

\item{reference, }{the sequence indicating the empty plasmid}
}
\value{
a gene expression matrix with empty cells
}
\description{
Empty selection step of the data analysis pipeline for the iPS2-seq methods from HEDGe lab.
}
\examples{

catcheR_10Xnocatch(group = "docker", folder = folder, expression.matrix = "matrix.csv", threshold = 10, sample = 1)

}
\author{
Maria Luisa Ratto, marialuisa.ratto [at] unito [dot] it, UNITO
}
