% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catcheR_cell_filtering.R
\name{catcheR_cell_filtering}
\alias{catcheR_cell_filtering}
\title{catcheR_cell_filtering}
\usage{
catcheR_cell_filtering(
  group = c("docker", "sudo"),
  folder,
  expression.matrix,
  UMI.count,
  percentage = 15
)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{folder, }{a character string indicating the path of the working folder containing the input files}

\item{expression.matrix, }{a character string indicating the filename of the gene expression matrix csv}

\item{UMI.count, }{integer threshold of percentage of UMIs supporting a UCI to consider the UCI valid. Manually set.}

\item{percentage, }{integer threshold of percentage of UMIs supporting a UCI over total UMIs supporting UCIs in the same cell, to consider the UCI valid. Suggested default is 15.}
}
\value{
a gene expression matrix where cell names contains annotations about the perturbation present in each cell + QC plots
}
\description{
Cell filtering step of the data analysis pipeline for the iPS2-seq methods from HEDGe lab.
}
\author{
Maria Luisa Ratto, marialuisa.ratto [at] unito [dot] it, UNITO
}
