% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catcheR_empty.R
\name{catcheR_empty}
\alias{catcheR_empty}
\title{catcheR_empty}
\usage{
catcheR_empty(group = c("docker", "sudo"), folder, expression.matrix)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{folder, }{a character string indicating the path of the working folder containing the input files}

\item{expression.matrix, }{a character string indicating the filename of the gene expression matrix csv}
}
\value{
a gene expression matrix with empty cells
}
\description{
Empty selection step of the data analysis pipeline for the iPS2-seq methods from HEDGe lab.
}
\author{
Maria Luisa Ratto, marialuisa.ratto [at] unito [dot] it, UNITO
}
