% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catcheR_filtercatch.R
\name{catcheR_filtercatch}
\alias{catcheR_filtercatch}
\title{catcheR_filtercatch}
\usage{
catcheR_filtercatch(
  group = c("docker", "sudo"),
  folder,
  expression.matrix,
  UMI.count,
  percentage = 15,
  ratio = 5,
  sample = 1
)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{folder, }{a character string indicating the path of the working folder containing the input files}

\item{expression.matrix, }{a character string indicating the filename of the gene expression matrix csv}

\item{UMI.count, }{integer threshold of percentage of UMIs supporting a UCI to consider the UCI valid. Manually set.}

\item{percentage, }{integer threshold of percentage of UMIs supporting a UCI over total UMIs supporting UCIs in the same cell, to consider the UCI valid. Suggested default is 15.}

\item{ratio, }{integer indicating the : the ratio between the number of UMIs supporting the most represent UCI and the number of UMIs supporting the second most represented UCI to consider accept the cell as a single integration cell.}

\item{sample, }{integer indicating the sample to analyze (between those aggregated with cell ranger aggr)}
}
\value{
a gene expression matrix where cell names contains annotations about the perturbation present in each cell + QC plots
}
\description{
Cell filtering step of the data analysis pipeline for the iPS2-seq methods from HEDGe lab.
}
\examples{

catcheR_filtercatch(group = "docker", folder = "/20tb/ratto/catcheR/test_CM5/", expression.matrix = "matrix.csv", UMI.count = 10, 15, sample = 1)

}
\author{
Maria Luisa Ratto, marialuisa.ratto [at] unito [dot] it, UNITO
}
