% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catcher_load.R
\name{catcheR_load}
\alias{catcheR_load}
\title{catcheR_load}
\usage{
catcheR_load(
  group = c("docker", "sudo"),
  folder,
  expression.matrix,
  control_genes,
  control_samples = "NotSpecified",
  replicates = "NotSpecified",
  sample_names,
  resolution = 8e-04,
  genes = "NotSpecified"
)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{folder, }{a character string indicating the path of the working folder containing the input files and results of previous analysis}

\item{expression.matrix, }{a character string indicating the filename of the annotated gene expression matrix csv resulting from catcher catch ("filtered_annotated_silencing_matrix_complete_all_samples.csv")}

\item{control_genes, }{string indicating the file name of a file listing the genes which are used as controls}

\item{control_samples, }{string indicating the file name of a file listing the samples which are used as controls (no TET samples)}

\item{replicates, }{string indicating the file name of a file listing the name of the replicates and their order associated to samples (used for batch correction)}

\item{sample_names, }{string indicating the file name of a csv file listing the samples and the associated name}

\item{resolution, }{numeric indicating the resolution to be used for Monocle clustering}

\item{genes, }{string indicating the file name of a file listing the genes of interest whose expressin will be plotted on the UMAP}
}
\value{
a gene expression matrix with empty cells
}
\description{
Data loading step of the data analysis pipeline for the iPS2-seq methods from HEDGe lab.
}
\examples{

catcheR_load(group="docker",folder="/3tb/data/ratto/aggr/test/", expression.matrix = "annotated_silencing_matrix_complete_all_samples.csv",control_genes =  "controls.txt",control_samples = "noTET.txt",replicates = "replicates.txt",sample_names = "samples.csv",resolution = 8e-4,genes = "genelist.txt")

}
\author{
Maria Luisa Ratto, marialuisa.ratto [at] unito [dot] it, UNITO
}
