% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catcheR_merge.R
\name{catcheR_merge}
\alias{catcheR_merge}
\title{catcheR_merge}
\usage{
catcheR_merge(group = c("docker", "sudo"), folder, samples = 2, empty = T)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{folder, }{a character string indicating the path of the working folder containing the input files}

\item{samples, }{integer indicating the number of different samples present in the sample (aggregated with cell ranger aggr)}

\item{empty, }{logical, are add empty cells to the matrix? nocatch must have been run. Default TRUE}
}
\value{
a gene expression matrix where cell names contains annotations about the perturbation present in each cell
}
\description{
Merge annotated matrices in a single one, after tuning catcheR parameters.
}
\examples{

catcheR_merge(group = "docker", folder = "/20tb/ratto/catcheR/test_CM5/", samples = 2)

}
\author{
Maria Luisa Ratto, marialuisa.ratto [at] unito [dot] it, UNITO
}
