% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catcheR_pseudotime.R
\name{catcheR_pseudotime}
\alias{catcheR_pseudotime}
\title{catcheR_pseudotime}
\usage{
catcheR_pseudotime(group = c("docker", "sudo"), folder, cds, pseudotime)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{folder, }{a character string indicating the path of the working folder containing the input files and results of previous analysis}

\item{cds, }{a character string indicating the filename of the annotated gene expression matrix csv resulting from catcher catch ("filtered_annotated_silencing_matrix_complete_all_samples.csv")}

\item{pseudotime, }{a character string indicating the pseudotime csv file}
}
\value{
statistics on pseudotime
}
\description{
Data loading step of the data analysis pipeline for the iPS2-seq methods from HEDGe lab.
}
\author{
Maria Luisa Ratto, marialuisa.ratto [at] unito [dot] it, UNITO
}
