% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catcheR_regression.R
\name{catcheR_regression}
\alias{catcheR_regression}
\title{catcheR_regression}
\usage{
catcheR_regression(group = c("docker", "sudo"), folder, cds)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{folder, }{a character string indicating the path of the working folder containing the input files and results of previous analysis}

\item{cds, }{a character string indicating the filename of the annotated gene expression matrix csv resulting from catcher catch ("filtered_annotated_silencing_matrix_complete_all_samples.csv")}
}
\value{
tables of genes correlated with a specific perturbation group
}
\description{
Finds genes whose expression is correlated with a specific perturbation group of cell. Part of the data analysis pipeline for the iPS2-seq methods from HEDGe lab.
}
\author{
Maria Luisa Ratto, marialuisa.ratto [at] unito [dot] it, UNITO
}
