% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sci_tomatrix.R
\name{catcheR_sci_tomatrix}
\alias{catcheR_sci_tomatrix}
\title{catcheR_sci_tomatrix}
\usage{
catcheR_sci_tomatrix(
  group = c("docker", "sudo"),
  folder,
  sample.name,
  UMI.cutoff
)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{folder, }{a character string indicating the path of the working folder containing the input files}

\item{sample.name, }{a character string indicating the name of the experiment}

\item{UMI.cutoff, }{integer indicating the minimum number of UMI per cell to consider the cell valid}
}
\value{
a gene expression matrix + QC plots and stats
}
\description{
From fastq to gene expression matrix for SCI data.
}
\author{
Maria Luisa Ratto, marialuisa.ratto [at] unito [dot] it, UNITO
}
