% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catcheR_scicatch.R
\name{catcheR_scicatch}
\alias{catcheR_scicatch}
\title{catcheR_scicatch}
\usage{
catcheR_scicatch(
  group = c("docker", "sudo"),
  folder,
  expression.matrix,
  reference = "GGCGCGTTCATCTGGGGGAGCCG",
  UCI.length = 6,
  threads = 2,
  percentage = 15,
  ratio = 5,
  mode = "bimodal",
  x = 100,
  y = 400
)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{folder, }{a character string indicating the path of the working folder containing the input files. Must contain a folder "fastq" with demultiplexed fastq files}

\item{expression.matrix, }{a character string indicating the filename of the gene expression matrix csv}

\item{reference, }{a character string indicating the sequence to identify reads containing barcodes. Should be found at at the beginning of read2. Use reverse complement!}

\item{threads, }{integer number of threads to be used for parallelization}

\item{percentage, }{integer threshold of percentage of UMIs supporting a UCI over total UMIs supporting UCIs in the same cell, to consider the UCI valid. Suggested default is 15.}

\item{ratio, }{select valid cells = only 1 true UCI and top UCI UMIs / second top UCI UMIs > ratio}

\item{mode, }{a character string. Two options: "bimodal" or "noise". To evaluate a threshold number of UMIs to consider a UCI valid there are 2 options: "bimodal" (default) which sets the threshold at the valley of the UMIxUCI distribution, or "noise", which sets the threshold at 1.35 * number of UCI supported by a single UMI.}

\item{x, }{an integer indicating the x limit to zoom the UMIxUCI plot. Default is 100}

\item{y, }{an integer indicating the y limit to zoom the UMIxUCI plot. Default is 400}

\item{UCI.legth, }{integer indicating the length of unique clonal identifier. Should be found after reference on read2}
}
\value{
a gene expression matrix where cell names contains annotations about the perturbation present in each cell + QC plots
}
\description{
Data analysis pipeline for the iPS2-seq methods from HEDGe lab, using combinatorial single cell indexing sequencing.
}
\author{
Maria Luisa Ratto, marialuisa.ratto [at] unito [dot] it, UNITO
}
