% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catcheR_scicatchQC.R
\name{catcheR_scicatchQC}
\alias{catcheR_scicatchQC}
\title{catcheR_scicatchQC}
\usage{
catcheR_scicatchQC(
  group = c("docker", "sudo"),
  folder,
  expression.matrix,
  reference = "GGCGCGTTCATCTGGGGGAGCCG",
  mode = "bimodal",
  x = 100,
  y = 400
)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{folder, }{a character string indicating the path of the working folder containing the input files}

\item{expression.matrix, }{a character string indicating the filename of the gene expression matrix csv}

\item{reference, }{a character string indicating the sequence to identify reads containing barcodes. Should be found at at the beginning of read2. Use reverse complement!}

\item{mode, }{a character string. Two options: "bimodal" or "noise". To evaluate a threshold number of UMIs to consider a UCI valid there are 2 options: "bimodal" (default) which sets the threshold at the valley of the UMIxUCI distribution, or "noise", which sets the threshold at 1.35 * number of UCI supported by a single UMI.}

\item{x, }{an integer indicating the x limit to zoom the UMIxUCI plot. Default is 100}

\item{y, }{an integer indicating the y limit to zoom the UMIxUCI plot. Default is 400}
}
\value{
QC plots and intermediate data
}
\description{
Explorative step of the data analysis pipeline for the iPS2-seq methods from HEDGe lab with SCI.
}
\author{
Maria Luisa Ratto, marialuisa.ratto [at] unito [dot] it, UNITO
}
