% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catcheR_scicsv.R
\name{catcheR_scicsv}
\alias{catcheR_scicsv}
\title{catcheR_scicsv}
\usage{
catcheR_scicsv(group = c("docker", "sudo"), folder)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{folder, }{a character string indicating the path of the working folder containing the input files}
}
\value{
a gene expression matrix + QC plots and stats
}
\description{
From fastq to gene expression matrix for SCI data.
}
\author{
Maria Luisa Ratto, marialuisa.ratto [at] unito [dot] it, UNITO
}
