% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catcheR_scinocatch.R
\name{catcheR_scinocatch}
\alias{catcheR_scinocatch}
\title{catcheR_scinocatch}
\usage{
catcheR_scinocatch(
  group = c("docker", "sudo"),
  folder,
  expression.matrix,
  threshold,
  reference = "TACGCGTTCATCTGGGGGAGCCG"
)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{folder, }{a character string indicating the path of the working folder containing the input files}

\item{expression.matrix, }{a character string indicating the filename of the gene expression matrix csv}

\item{threshold, }{an integer indicating the minimum number of UMIs associated to the empty reference to consider a cell empty}

\item{reference, }{the sequence indicating the empty plasmid}
}
\value{
a gene expression matrix with empty cells
}
\description{
Empty selection step of the data analysis pipeline for the iPS2-seq methods from HEDGe lab.
}
\author{
Maria Luisa Ratto, marialuisa.ratto [at] unito [dot] it, UNITO
}
