% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catcheR_sortcatch.R
\name{catcheR_sortcatch}
\alias{catcheR_sortcatch}
\title{catcheR_sortcatch}
\usage{
catcheR_sortcatch(
  group = c("docker", "sudo"),
  folder,
  expression.matrix,
  swaps
)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{folder, }{a character string indicating the path of the working folder containing the input files}

\item{expression.matrix, }{a character string indicating the filename of the csv file of the annotated gene expression produced by catcheR_10Xcatch or catcheR_scicatch i.e. silencing_matrix.csv}

\item{swaps, }{a character string indicating the filename of the txt file indicating swaps to be corrected, output of catcheR_step1QC, i.e. ”reliable_clones_swaps.txt"}
}
\value{
Updated gene expression matrix called ”silencing_matrix_updated.csv"
}
\description{
Modify annotated gene expression matrix from catcheR_10Xcatch or catcheR_scicatch based on the barcode swaps identified with catcheR_step1QC.
}
\author{
Maria Luisa Ratto, marialuisa.ratto [at] unito [dot] it, UNITO
}
