% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catcheR_step1QC.R
\name{catcheR_step1QC}
\alias{catcheR_step1QC}
\title{catcheR_step1QC}
\usage{
catcheR_step1QC(
  group = c("docker", "sudo"),
  folder,
  fastq.read1,
  DIs = 100,
  ratio = 10,
  plot.threshold = 2000,
  clones = NULL
)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{folder, }{a character string indicating the path of the working folder containing the input files}

\item{fastq.read1, }{a character string indicating the filename of read 1 fastq or fastq.gz containing barcodes sequencing}

\item{DIs, }{integer of the minimum number of diversity indexes (DIs, pseudo-unique reads) of the most represented shRNA matching to a given UCI-BC; in combination with "ratio", it selects UCI-BCs for which it is possible to reliably assign an shRNA. Default is 100}

\item{ratio, }{integer of the minimum ratio between the number of DIs of the most represented and second most represented shRNAs matching to a given UCI-BC. in combination with DI, it is employed to filter UCI-BCs for which it is possible to reliably assign an shRNA. Default is 10}

\item{plot.threshold, }{integer of the minimum number of DIs per UCI-BC for output plot}

\item{clones, }{a character string indicating the filename of txt file containing a newline separated list of clones of interest in the format of barcode_UCI}
}
\value{
plots and stats
}
\description{
For the analysis of the plasmids in their final form, i.e. where only the barcode is present
}
\author{
Maria Luisa Ratto, marialuisa.ratto [at] unito [dot] it, UNITO
}
