% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catcheR_step2QC.R
\name{catcheR_step2QC}
\alias{catcheR_step2QC}
\title{catcheR_step2QC}
\usage{
catcheR_step2QC(
  group = c("docker", "sudo"),
  folder,
  fastq.read1,
  fastq.read2,
  threshold.percentage = 300,
  plot.threshold = 2000,
  clones = NULL
)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{folder, }{a character string indicating the path of the working folder containing the input files}

\item{fastq.read1, }{a character string indicating the filename of read 1 fastq or fastq.gz containing barcodes sequencing}

\item{fastq.read2, }{a character string indicating the filename of read 2 fastq or fastq.gz containing barcodes sequencing}

\item{threshold.percentage, }{integer, a minimum of reads associated to a clone to show it in plots of percentage of reads per clone, for visualization purposes.}

\item{plot.threshold, }{integer, a minimum of reads associated to a clone to show it in plots of reads per clone, for visualization purposes.}

\item{clones, }{a character string indicating the filename of txt file containing a newline separated list of clones of interest in the format of barcode_UCI}
}
\value{
plots and stats
}
\description{
For the analysis of the plasmids in their final form, i.e. where only the barcode is present
}
\author{
Maria Luisa Ratto, marialuisa.ratto [at] unito [dot] it, UNITO
}
