% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_op.R
\name{delete_cache}
\alias{delete_cache}
\title{Delete your entire cache}
\usage{
delete_cache(really = FALSE)
}
\arguments{
\item{really}{set this to TRUE if you actually want to delete}
}
\value{
an empty cache folder
}
\description{
Delete your entire cache
}
\examples{
\dontrun{delete_cache(TRUE)}
}

