% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_utils.R
\name{exists_in_cache}
\alias{exists_in_cache}
\title{Check if given key exists in cache}
\usage{
exists_in_cache(key, max_lifetime)
}
\arguments{
\item{key}{hashed representation of function and args}

\item{max_lifetime}{max age of cached object (days)}
}
\value{
boolean
}
\description{
Check if given key exists in cache
}
\examples{
exists_in_cache("de245179163e5245a56484e7207bf3a3469c358b")
}

