% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_utils.R
\name{hash_query}
\alias{hash_query}
\title{Hash an arbitrary string}
\usage{
hash_query(query)
}
\arguments{
\item{query}{to hash}
}
\value{
sha1 hash
}
\description{
Hash an arbitrary string
}
\examples{
hash_query("sin1234")
}

