% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_utils.R
\name{match_fun}
\alias{match_fun}
\title{Smart matching of functions}
\usage{
match_fun(fun)
}
\arguments{
\item{fun}{function name, as a string}
}
\value{
the desired function
}
\description{
Smart matching of functions
}
\examples{
catcher:::match_fun("digest::digest")
catcher:::match_fun("sum")
}

