% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oligo_design.R
\name{oligo_design}
\alias{oligo_design}
\title{oligo_design}
\usage{
oligo_design(
  group = c("docker", "sudo"),
  folder,
  sequences,
  gibson.five = "AGTTCCCTATCAGTGATAGAGATCCC",
  gibson.three = "GTAGCTCGCTGATCAGC",
  fixed = "GTCGACATTTAAATGGCGCGCC",
  restriction.sites
)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{folder, }{a character string indicating the path of the working folder containing the input files}

\item{sequences, }{a character string indicating the filename of the csv file containing sequences of interest and corresponding barcodes}

\item{gibson.five, }{a character string indicating the sequence of the gibson overhang to put before the shRNA, at the beginning of the construct}

\item{gibson.three, }{a character string indicating the sequence of the gibson overhang to put at the end of the construct}

\item{fixed, }{a character string indicating the fixed sequence to put after the shRNA in the contruct, between poly T and UCI}

\item{restriction.sites, }{a character string indicating the filename of the txt file containing sequences of the restriction sites to avoid in the sequence}
}
\value{
a file called "output.txt" containing the sequences of the constructs. Restriction sites are in lower case.
}
\description{
Fucntion to design oligos for iPS2-seq methods from HEDGe lab.
}
\author{
Maria Luisa Ratto, marialuisa.ratto [at] unito [dot] it, UNITO
}
