% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_utils.R
\name{read_from_cache}
\alias{read_from_cache}
\title{Read file from cache.}
\usage{
read_from_cache(key)
}
\arguments{
\item{key}{hashed representation of object to look up}
}
\value{
cached file
}
\description{
Also includes when the file was created and how old it is.
}
\examples{
\dontrun{read_from_cache("6904e1dbc962a5df3040efd454ade0d564be27ce")}
}

