% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache-utils.R
\name{save_to_cache}
\alias{save_to_cache}
\title{Save R object to cache}
\usage{
save_to_cache(df, key)
}
\arguments{
\item{df}{arbitary R object}

\item{key}{hashed representation of df's generating query}
}
\value{
invisibly returns df
}
\description{
Save R object to cache
}
\examples{
\dontrun{save_to_cache(ggplot2::diamonds, "6904e1dbc962a5df3040efd454ade0d564be27ce")}
}

