% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_op.R
\name{cache_op}
\alias{cache_op}
\title{Wrapper to add caching to functions}
\usage{
cache_op(fun, query, use_cache = TRUE, overwrite = FALSE,
  max_lifetime = 30, ...)
}
\arguments{
\item{fun}{function name (string)}

\item{query}{first argument to fun}

\item{use_cache}{should we use cache this time}

\item{overwrite}{if cached file exists, should the function run and overwrite?}

\item{max_lifetime}{max acceptable age of a cached object}

\item{...}{other parameters passed to fun}
}
\value{
typical fun, query, ellipsis result
}
\description{
Wrapper to add caching to functions
}
\examples{
cache_op("sin", pi)
# note: this below only works on windows; just an example of additional params
cache_op("read.csv", "https://cdn.rawgit.com/Keno/8573181/raw/7e97f56f521d1f49b966e04457687e87da1b062b/gistfile1.txt", header = TRUE)
}

