% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catcheR_enrichment.R
\name{catcheR_enrichment}
\alias{catcheR_enrichment}
\title{catcheR_enrichment}
\usage{
catcheR_enrichment(
  group = c("docker", "sudo"),
  folder,
  file,
  meta,
  timepoint = "PSC",
  control_gene = "SCR",
  min_cells_cluster = 70,
  min_cells_shRNA = 40
)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{folder, }{a character string indicating the path of the working folder containing the input files and results of previous analysis}

\item{file, }{a character string indicating the filename of the processed cds file resulting from catcheR_load}

\item{meta, }{string indicating the file name of the metadata file resulting from catcheR_load}

\item{timepoint, }{string indicating the starting timepoint to compare other timepoints to (e.g. day 0) (optional)}

\item{control_gene, }{string indicating the control gene to calculate statistics against in case internal control (no TET) is not present (e.g. "SCR") (optional)}

\item{min_cells_cluster, }{integer indicating the threshold for minimum number of cells per cluster to keep the cluster (optional) default: 70}

\item{min_cells_shRNA, }{integer indicating the threshold for minimum number of cells per shRNA to keep the cluster (optional) default: 40}
}
\value{
plots and pvalues
}
\description{
Data loading step of the data analysis pipeline for the iPS2-seq methods from HEDGe lab.
}
\examples{

catcheR_enrichment(group = "docker", folder = folder, file = "processed_cds.RData", meta = "cell_metadata.csv", timepoint = "PSC", control_gene = "SCR")

}
\author{
Maria Luisa Ratto, marialuisa.ratto [at] unito [dot] it, UNITO
}
